feat: add user auth to todo sub route

+++ internal/database/db.go
@@ -24,7 +24,7 @@ func Initialize() {
 func createTables() {
 	userTable := `
     CREATE TABLE IF NOT EXISTS users (
-        id INTEGER PRIMARY KEY AUTOINCREMENT,
+        id FLOAT PRIMARY KEY AUTOINCREMENT,
         name TEXT NOT NULL,
         email TEXT UNIQUE NOT NULL,
         password TEXT NOT NULL
